Remove generated documentation
Including generated documentation files in the source repository makes
it more difficult for contributors to offer documentation patches,
since they effectively need to "build" the docs for every change,
which may require toolsets that aren't installed by default and with
which contributors may not be familiar.

This commit removes the generated documentation from the repository.
These files are already rebuilt automatically by the release scripts.

Signed-off-by: Lars Kellogg-Stedman <lars@oddbit.com>